requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
Pillow==10.0.0
Jinja2==3.1.3
python-multipart==0.0.9 